Moved all constants to data.js in constants directory, removed exercisesArray.js, updated imports
