+++ src/components/TabContent/TabContent.jsx
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import Section from "../Section.jsx";
 import Tabs from "../Tabs/Tabs.jsx";
 import Card from "../UI/Card.jsx";
-import { PARTS_OF_PLANS } from "../../data.js";
 import TabButton from "../TabButton/TabButton.jsx";
 import EditButton from "../EditButton/EditButton.jsx";
+import {PARTS_OF_PLANS} from "../../constants/data.js";
 
 const TabContent = () => {
   const [selectedPartOfPlan, setSelectedPartOfPlan] = useState();
