+++ src/constants/data.js
@@ -5,3 +5,37 @@ export const INITIAL_EXERCISES_INPUT = {
     tempo: "",
     rest: "",
 };
+
+export const PLANS = [
+    {
+        image: "",
+        title: "Plan Kasi",
+    },
+    {
+        image: "",
+        title: "Plan Kuby",
+    },
+    {
+        image: "",
+        title: "Plan Damiana",
+    },
+    {
+        image: "",
+        title: "Plan Uli",
+    },
+];
+
+export const PARTS_OF_PLANS = [
+    {
+        image: "",
+        title: "Rozgrzewka",
+    },
+    {
+        image: "",
+        title: "Trening A",
+    },
+    {
+        image: "",
+        title: "Trening B",
+    },
+];
