--- src/data.js
@@ -1,33 +1,0 @@
-export const PLANS = [
-  {
-    image: "",
-    title: "Plan Kasi",
-  },
-  {
-    image: "",
-    title: "Plan Kuby",
-  },
-  {
-    image: "",
-    title: "Plan Damiana",
-  },
-  {
-    image: "",
-    title: "Plan Uli",
-  },
-];
-
-export const PARTS_OF_PLANS = [
-  {
-    image: "",
-    title: "Rozgrzewka",
-  },
-  {
-    image: "",
-    title: "Trening A",
-  },
-  {
-    image: "",
-    title: "Trening B",
-  },
-];
