--- src/utils/exercisesArray.js
@@ -1,19 +1,0 @@
-export function createArrayOfExercises({
-  order,
-  exercise,
-  seriesAndRepetitions,
-  tempo,
-  rest,
-}) {
-  const exercisesArr = [];
-
-  exercisesArr.push({
-    order,
-    exercise,
-    seriesAndRepetitions,
-    tempo,
-    rest,
-  });
-
-  return exercisesArr;
-}
